Add Listeners and Parameter injection
Added the ListenersGroupTestDemo.xml and the sauceDemo.xlm files to
the project.

SauceDemo.xml is where we have added and defined the parameters that
will be used in the SauceDemoLogin Test class.
@Parameters annotaion has been added to the methods which will accept
the injected parameters from the .xml file.

ListenersGroupTestDemo.xml file has added the <listeners> tags where
a developer will add the <listener> tags inside of them and define the
package.class from which the Listeners will come from.
Example: <listener class-name="common.Listeners />