Add Screenshot capabilities
BaseTest to run setup and teardown before test suite.
Test Utilities extends base test and has the getScreenShot function
which takes a screnshot and writes it to the screenshots folder.

ListenersDemoTest just takes a screenshot of ebay landing pay.
Test designed to fail with an assert=false.

ListenersDemoTest.xml is the .xml for this test