Add ReportNG to the project
ReportNG maven repository has been added to the project.  This will
make the reports much easier to read and understand.
Adjustments have been made for the addition of the ReportNG module,
such as fixing the bug that stopped the report from displaying
the link to the failing test screenshot.
The .xml file has been changed to add the reportNG listener.